Choose between EntraID and HID, depending on env & config
On local environments, use HID.

On non-local, use EntraID if configured,
otherwise use HID
