+++ libs/hpc-live/src/lib/hid-live.ts
@@ -1,0 +1,147 @@
+import type { Session } from '@unocha/hpc-core';
+import {
+  UserManager,
+  WebStorageStateStore,
+  type OidcMetadata,
+  type User,
+} from 'oidc-client';
+import type { AuthProvider, AuthResult, HIDConfig } from './interfaces';
+import { LiveModel } from './model';
+
+/**
+ * Construct the OIDC Metadata for HID,
+ * this is necessary because HID doesn't include information about the
+ * logout endpoint in its metadata, and we need to add it manually.
+ */
+const getOpenIDMetadata = async (authUrl: string): Promise<OidcMetadata> => {
+  const metadataEndpoint = new URL(
+    '/.well-known/openid-configuration',
+    authUrl
+  );
+  const logoutEndpoint = new URL('/logout', authUrl);
+  logoutEndpoint.searchParams.set('redirect', location.origin);
+
+  const req = await fetch(metadataEndpoint.href);
+  if (!req.ok) {
+    throw new Error('Unable to get HID metadata');
+  }
+  const metadata: OidcMetadata = await req.json();
+  metadata.end_session_endpoint = logoutEndpoint.href;
+  return metadata;
+};
+
+export class HIDProvider implements AuthProvider {
+  private readonly config: HIDConfig;
+  private readonly store: WebStorageStateStore;
+  private readonly userManager: Promise<UserManager>;
+
+  public constructor(config: HIDConfig) {
+    this.config = config;
+
+    const redirectUri = new URL(document.location.href);
+    redirectUri.pathname = '/';
+    redirectUri.hash = '';
+
+    this.store = new WebStorageStateStore({
+      store: localStorage,
+    });
+
+    (
+      globalThis as unknown as { clearSessionStorage: () => Promise<void> }
+    ).clearSessionStorage = this.clearSessionStorage;
+
+    this.userManager = getOpenIDMetadata(config.hpcAuthUrl).then(
+      (metadata) =>
+        new UserManager({
+          client_id: config.hpcAuthClientId,
+          authority: config.hpcAuthUrl,
+          redirect_uri: redirectUri.href,
+          response_type: 'token',
+          scope: 'profile',
+          userStore: this.store,
+          metadata,
+        })
+    );
+  }
+
+  public clearSessionStorage = async (): Promise<void> => {
+    const keys = await this.store.getAllKeys();
+    await Promise.all(keys.map((key) => this.store.remove(key)));
+    globalThis.location.reload();
+  };
+
+  private getSessionUser = async (
+    user: User | null
+  ): Promise<Session['getUser']> => {
+    if (!user) {
+      return () => null;
+    }
+
+    const accountUrl = new URL('/account.json', this.config.hpcAuthUrl);
+    const res = await fetch(accountUrl.href, {
+      headers: {
+        Authorization: `Bearer ${user.access_token}`,
+      },
+    });
+
+    if (!res.ok) {
+      return () => ({
+        name: 'unknown',
+      });
+    }
+
+    const info = await res.json();
+    return () => ({
+      name: info.name ?? 'unknown',
+    });
+  };
+
+  public init = async (): Promise<AuthResult> => {
+    const userManager = await this.userManager;
+
+    try {
+      const user = await userManager.signinRedirectCallback();
+      const redirectTo = user.state ?? document.location.pathname;
+      if (history.replaceState) {
+        history.replaceState(null, document.title, redirectTo);
+        // TODO: interact directly with React Router history to get it to reload
+        // the route without needing to reload the page
+        globalThis.location.reload();
+      } else {
+        globalThis.location = redirectTo;
+      }
+    } catch {
+      // No sign in response, that's fine
+    }
+
+    const user = await userManager.getUser();
+
+    const session: Session = {
+      getUser: await this.getSessionUser(user),
+      logIn: () =>
+        userManager.signinRedirect({
+          state: globalThis.location.href,
+        }),
+      logOut: () =>
+        userManager.signoutRedirect().then(() => userManager.removeUser()),
+    };
+
+    // When user logs in/out in a different tab, log in/out in current tab as well
+    globalThis.addEventListener('storage', (e) => {
+      // This is how oidc-client creates its storage keys
+      const keyPart = `user:${this.config.hpcAuthUrl}:${this.config.hpcAuthClientId}`;
+      if (e.key?.indexOf(keyPart) !== -1) {
+        // Reload the window to have new session from different tab applied to current one
+        globalThis.location.reload();
+      }
+    });
+
+    const model = new LiveModel({
+      baseUrl: this.config.hpcApiUrl,
+      hidToken: user?.access_token ?? null,
+      clearSessionStorage: this.clearSessionStorage,
+    });
+
+    return { session, model };
+  };
+}
