+++ libs/hpc-live/src/lib/live.ts
@@ -1,161 +1,77 @@
+import { type config } from '@unocha/hpc-core';
+import { EntraIDProvider } from './entraid-live';
+import { HIDProvider } from './hid-live';
 import {
-  UserManager,
-  WebStorageStateStore,
-  type OidcMetadata,
-  type User,
-} from 'oidc-client';
-
-import { type config, type Session } from '@unocha/hpc-core';
-
-import { LiveModel } from './model';
-
-/**
- * Construct the OIDC Metadata for HID,
- * this is necessary because HID doesn't include information about the
- * logout endpoint in its metadata, and we need to add it manually.
- */
-const getOpenIDMetadata = async (authUrl: string): Promise<OidcMetadata> => {
-  const metadataEndpoint = new URL(
-    '/.well-known/openid-configuration',
-    authUrl
-  );
-  const logoutEndpoint = new URL('/logout', authUrl);
-  logoutEndpoint.searchParams.set('redirect', location.origin);
-
-  const req = await fetch(metadataEndpoint.href);
-  if (!req.ok) {
-    throw new Error('Unable to get HID metadata');
-  }
-  const metadata: OidcMetadata = await req.json();
-  metadata.end_session_endpoint = logoutEndpoint.href;
-  console.log(metadata);
-  return metadata;
-};
+  AuthProviderType,
+  type AuthProvider,
+  type AuthResult,
+  type EntraIDConfig,
+} from './interfaces';
 
 export class LiveBrowserClient {
-  private readonly config: config.Config;
-  private readonly store: WebStorageStateStore;
-  private readonly userManager: Promise<UserManager>;
+  private readonly authProvider: AuthProvider;
 
   public constructor(config: config.Config) {
-    console.log('initializing live', config);
+    const providerType = this.determineProviderType(config);
 
-    this.config = config;
+    switch (providerType) {
+      case AuthProviderType.ENTRA_ID:
+        // Even though we determined to use Entra ID,
+        // TS cannot infer that `entraClientId` and `entraTenantId` are defined
+        if (!this.hasEntraIDConfig(config)) {
+          throw new Error('Missing Entra ID configuration');
+        }
 
-    const redirectUri = new URL(document.location.href);
-    redirectUri.pathname = '/';
-    redirectUri.hash = '';
+        this.authProvider = new EntraIDProvider(config);
+        break;
+      case AuthProviderType.HID:
+        this.authProvider = new HIDProvider(config);
+        break;
+      default:
+        throw new Error(`Unsupported auth provider type: ${providerType}`);
+    }
+  }
 
-    this.store = new WebStorageStateStore({
-      store: localStorage,
-    });
+  /**
+   * Initialize authentication and return session + model
+   */
+  public init = (): Promise<AuthResult> => {
+    return this.authProvider.init();
+  };
 
-    (
-      globalThis as unknown as { clearSessionStorage: () => Promise<void> }
-    ).clearSessionStorage = this.clearSessionStorage;
+  /**
+   * Clear session storage
+   */
+  public clearSessionStorage = (): void => {
+    this.authProvider.clearSessionStorage();
+  };
 
-    this.userManager = getOpenIDMetadata(config.hpcAuthUrl).then(
-      (metadata) =>
-        new UserManager({
-          client_id: config.hpcAuthClientId,
-          authority: config.hpcAuthUrl,
-          redirect_uri: redirectUri.href,
-          response_type: 'token',
-          scope: 'profile',
-          userStore: this.store,
-          metadata,
-        })
+  /**
+   * Checks if we're running in local environment
+   */
+  private isLocal(): boolean {
+    return (
+      ['localhost', '127.0.0.1'].includes(globalThis.location.hostname) &&
+      globalThis.location.port === '4200' // Default port for Nx Webpack dev server
     );
   }
 
-  private clearSessionStorage = async () => {
-    const keys = await this.store.getAllKeys();
-    await Promise.all(keys.map((key) => this.store.remove(key)));
-    globalThis.location.reload();
-  };
+  /**
+   * Checks if Entra ID configuration is present and valid
+   */
+  private hasEntraIDConfig(config: config.Config): config is EntraIDConfig {
+    return !!(config.entraClientId && config.entraTenantId);
+  }
 
-  private getSessionUser = async (
-    user: User | null
-  ): Promise<Session['getUser']> => {
-    if (!user) {
-      return () => null;
+  /**
+   * Determines which auth provider to use based on configuration and environment
+   */
+  private determineProviderType(config: config.Config): AuthProviderType {
+    // In non-local environments, prefer Entra ID, if configured
+    if (!this.isLocal() && this.hasEntraIDConfig(config)) {
+      return AuthProviderType.ENTRA_ID;
     }
-    const accountUrl = new URL('/account.json', this.config.hpcAuthUrl);
-    const res = await fetch(accountUrl.href, {
-      headers: {
-        Authorization: `Bearer ${user.access_token}`,
-      },
-    });
-    if (!res.ok) {
-      return () => ({
-        name: 'unknown',
-      });
-    }
-    const info = await res.json();
-    return () => ({
-      name: info.name ?? 'unknown',
-    });
-  };
 
-  public init = async () => {
-    const userManager = await this.userManager;
-    // Attempt to load
-    await userManager
-      .signinRedirectCallback()
-      .then((user) => {
-        const redirectTo = user.state ?? document.location.pathname;
-        if (history.replaceState) {
-          history.replaceState(null, document.title, redirectTo);
-          // TODO: interact directly with React Router history to get it to reload
-          // the route without needing to reload the page
-          globalThis.location.reload();
-        } else {
-          globalThis.location = redirectTo;
-        }
-      })
-      .catch(() => {
-        // No sign in response, that's fine
-      });
-    const user = await userManager.getUser();
-    const session: Session = {
-      getUser: await this.getSessionUser(user),
-      logIn: () =>
-        userManager.signinRedirect({
-          state: globalThis.location.href,
-        }),
-      logOut: () =>
-        userManager.signoutRedirect().then(() => userManager.removeUser()),
-    };
-
-    // When user logs in/out in a different tab, log in/out in current tab as well
-    globalThis.addEventListener('storage', (e) => {
-      // This is how oidc-client creates its storage keys
-      const keyPart = `user:${this.config.hpcAuthUrl}:${this.config.hpcAuthClientId}`;
-      if (e.key?.indexOf(keyPart) !== -1) {
-        // Reload the window to have new session from different tab applied to current one
-        globalThis.location.reload();
-      }
-    });
-
-    if (user) {
-      const result = {
-        session,
-        model: new LiveModel({
-          baseUrl: this.config.hpcApiUrl,
-          hidToken: user.access_token,
-          clearSessionStorage: this.clearSessionStorage,
-        }),
-      };
-      return result;
-    }
-    const result = {
-      session,
-      model: new LiveModel({
-        baseUrl: this.config.hpcApiUrl,
-        hidToken: null,
-        clearSessionStorage: this.clearSessionStorage,
-      }),
-    };
-    return result;
-  };
+    return AuthProviderType.HID;
+  }
 }
